relocate & rename RolesEnum
