+++ src/enum/RolesEnum.js
@@ -1,0 +1,6 @@
+const RolesEnum = {
+    ADMIN: "admin",
+    USER: "user",
+};
+
+module.exports = RolesEnum;
